requirements.txt
# Deep Learning Core
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Scientific Computing
numpy>=1.24.3
scipy>=1.10.1
scikit-learn>=1.2.2
pandas>=2.0.3

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
tabulate>=0.8.9
pandas>=2.0.3

# System Monitoring
psutil>=5.9.5
nvidia-smi>=0.1.3
gpustat>=1.1.0
py3nvml>=0.2.7

# Data Processing
h5py>=3.9.0
pillow>=10.0.0

# Development Tools
tqdm>=4.65.0
jupyter>=1.0.0
ipython>=8.14.0
pyyaml>=6.0.1

# Code Quality
black>=23.7.0
flake8>=6.1.0
pytest>=7.4.0

# Optional Dependencies
scikit-image>=0.21.0

# Data generation
fenics
git+https://bitbucket.org/fenics-project/mshr

# For CUDA 11.8 compatibility, use:
--extra-index-url https://download.pytorch.org/whl/cu118
